scale-test: use multiple instance-groups
GCE supports a maximum of 1000 nodes in a managed instance group. Due to
this limitation, we need to use additional instance-groups for clusters
larger than 1000 nodes.

The forked create-instance-group action should be reverted once the
changes are merged into main. This was done to decrease the size of the
root volumes.

Signed-off-by: Tim Horner <timothy.horner@isovalent.com>