ci: use `aks-preview` in AKS CI when necessary
Due to AKS K8s version policy, the AKS Preview Extension[1] is required
in order to set up AKS clusters using the latest version of K8s as soon
as possible.

However, adding this extension to the CI workflow creates a discrepancy
compared to how we recommend and expect users to deploy Cilium on AKS
clusters. For example there have been instances where new AKS clusters
would not have the same defaults with and without the preview extension.

In order to minimize drift between the AKS workflow and documentation,
we introduce a new flag to the AKS K8s matrix configuration, allowing to
specify when the extension should be installed and thereby also allowing
deployment of AKS clusters using the latest version of K8s while it is
still in preview mode on AKS.

[1]: https://github.com/Azure/azure-cli-extensions/tree/main/src/aks-preview

Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>